third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-value-selectedOption.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Test that slotted-in selected-value part is updated to value of select
  assert_equals: Custom selected value part should be set to initial value of select expected "one" but got "Default custom selected-value text"
[FAIL] Test that option parts in a slotted-in listbox are reflected in the value property
  assert_equals: Custom selected value part should be set to initial value of select expected "one" but got "Default custom selected-value text"
Harness: the test ran to completion.
